Footsteps Progress New Bug: Footsteps GFX Crash

+++ External/NDYGFX/Include/Vector.h
@@ -128,7 +128,7 @@ public:
 
 	inline float GetLength() const
 	{
-		return sqrt(pow(this->x, 2) + pow(this->y, 2));
+		return sqrtf(powf(this->x, 2.0f) + powf(this->y, 2.0f));
 	}
 
 	inline void Normalize()
